requirements.txt
pandas
numpy
scikit-learn==1.0.2
scikit-surprise==1.1.3
matplotlib==3.5.3 